// src/types/voice.ts
export interface Voice {
  accent: string;
  age: string;
  gender: string;
  hq: boolean;
  isExperimental: boolean;
  isNew: boolean;
  isPopular: boolean;
  language: string;
  languageCode: string;
  loudness: string;
  name: string;
  sample: string;
  style: string;
  tempo: string;
  texture: string;
  value: string;
  voiceId: string;
}